CICD PIPELINE  FOR WEBSITE DEPLOYMENT
I developed a CI/CD pipeline project to automate website deployment. In this project, I utilized more than six DevOps tools, including AWS Cloud services, to create an efficient and scalable deployment process. The goal was to automate the deployment of a website to the server, ensuring faster and more reliable releases. Additionally, I focused on enhancing the security of the deployment process, ensuring that each stage of the pipeline was secured and compliant with best practices